run async loop
