+++ src/litserve/loops/__init__.py
@@ -1,0 +1,9 @@
+import multiprocessing as mp
+
+from litserve.loops.base import LitLoop, get_default_loop, inference_worker
+from litserve.loops.loops import ContinuousBatchingLoop, Output
+
+mp.allow_connection_pickling()
+
+
+__all__ = ["ContinuousBatchingLoop", "LitLoop", "get_default_loop", "inference_worker", "Output"]
